+++ src/App.jsx
@@ -7,6 +7,8 @@ import PageHeader from './components/Layout/PageHeader/PageHeader'
 import PublishModal from './components/PublishModal/PublishModal'
 import LaunchModal from './components/LaunchModal/LaunchModal'
 import LaunchImageModal from './components/LaunchModal/LaunchImageModal'
+import UploadGeojsonModal from './components/UploadGeojsonModal/UploadGeojsonModal'
+import SystemMessage from './components/SystemMessage/SystemMessage'
 
 import { GetCollectionsService } from './services/get-collections-service'
 import { useSelector } from 'react-redux'
@@ -21,6 +23,12 @@ function App() {
   const _showLaunchImageModal = useSelector(
     (state) => state.mainSlice.showLaunchImageModal
   )
+  const _showUploadGeojsonModal = useSelector(
+    (state) => state.mainSlice.showUploadGeojsonModal
+  )
+  const _showApplicationAlert = useSelector(
+    (state) => state.mainSlice.showApplicationAlert
+  )
   useEffect(() => {
     GetCollectionsService()
   }, [])
@@ -33,6 +41,10 @@ function App() {
         {_showPublishModal ? <PublishModal /> : null}
         {_showLaunchModal ? <LaunchModal /> : null}
         {_showLaunchImageModal ? <LaunchImageModal /> : null}
+        {_showUploadGeojsonModal ? (
+          <UploadGeojsonModal></UploadGeojsonModal>
+        ) : null}
+        {_showApplicationAlert ? <SystemMessage></SystemMessage> : null}
       </div>
     </React.StrictMode>
   )