+++ src/redux/slices/mainSlice.js
@@ -31,7 +31,11 @@ const initialState = {
   showAdvancedSearchOptions: false,
   isDrawingEnabled: false,
   mapDrawPolygonHandler: null,
-  searchGeojsonBoundary: null
+  searchGeojsonBoundary: null,
+  showUploadGeojsonModal: false,
+  showApplicationAlert: false,
+  applicationAlertMessage: 'System Error',
+  applicationAlertSeverity: 'error'
 }
 
 // next, for every key in the initialState
@@ -123,6 +127,18 @@ export const mainSlice = createSlice({
     },
     setsearchGeojsonBoundary: (state, action) => {
       state.searchGeojsonBoundary = action.payload
+    },
+    setshowUploadGeojsonModal: (state, action) => {
+      state.showUploadGeojsonModal = action.payload
+    },
+    setshowApplicationAlert: (state, action) => {
+      state.showApplicationAlert = action.payload
+    },
+    setapplicationAlertMessage: (state, action) => {
+      state.applicationAlertMessage = action.payload
+    },
+    setapplicationAlertSeverity: (state, action) => {
+      state.applicationAlertSeverity = action.payload
     }
   }
 })
@@ -157,5 +173,9 @@ export const { setshowAdvancedSearchOptions } = mainSlice.actions
 export const { setisDrawingEnabled } = mainSlice.actions
 export const { setmapDrawPolygonHandler } = mainSlice.actions
 export const { setsearchGeojsonBoundary } = mainSlice.actions
+export const { setshowUploadGeojsonModal } = mainSlice.actions
+export const { setshowApplicationAlert } = mainSlice.actions
+export const { setapplicationAlertMessage } = mainSlice.actions
+export const { setapplicationAlertSeverity } = mainSlice.actions
 
 export default mainSlice.reducer