+++ src/utils/mapHelper.js
@@ -19,6 +19,7 @@ import {
   VITE_MOSAIC_TILER_PARAMS
 } from '../assets/config'
 import { GetMosaicBoundsService } from '../services/get-mosaic-bounds'
+import GeoJSONValidation from './geojsonValidation'
 
 export const footprintLayerStyle = {
   color: '#3183f5',
@@ -46,7 +47,24 @@ export const clickedFootprintLayerStyle = {
   pane: 'searchResults'
 }
 
-export const customDrawingPolygonStyle = {
+const customSearchPointIconStyle = L.icon({
+  iconSize: [25, 41],
+  iconAnchor: [10, 41],
+  popupAnchor: [2, -40],
+  iconUrl: '/marker-icon.png',
+  shadowUrl: '/marker-shadow.png'
+})
+
+export const customSearchLineStyle = {
+  color: '#00C07B',
+  weight: 2,
+  opacity: 1,
+  dashArray: '4, 4',
+  dashOffset: '0',
+  pane: 'drawPane'
+}
+
+export const customSearchPolygonStyle = {
   color: '#00C07B',
   weight: 2,
   opacity: 1,
@@ -517,7 +535,7 @@ export function enableMapPolyDrawing() {
       map.eachLayer(function (layer) {
         if (layer.layer_name === 'drawBoundsLayer') {
           const drawLayer = e.layer
-          drawLayer.setStyle(customDrawingPolygonStyle)
+          drawLayer.setStyle(customSearchPolygonStyle)
           drawLayer.options.interactive = false
           layer.addLayer(drawLayer)
           const data = layer.toGeoJSON()
@@ -533,5 +551,76 @@ export function disableMapPolyDrawing() {
   const map = store.getState().mainSlice.map
   if (map && Object.keys(map).length > 0) {
     store.getState().mainSlice.mapDrawPolygonHandler.disable()
+  }
+}
+
+export function addUploadedGeojsonToMap(geojson) {
+  const map = store.getState().mainSlice.map
+  if (map && Object.keys(map).length > 0) {
+    clearLayer('drawBoundsLayer')
+    map.eachLayer(function (layer) {
+      if (layer.layer_name === 'drawBoundsLayer') {
+        let geojsonLayer = L.geoJSON(geojson)
+
+        geojsonLayer = L.geoJSON(geojson, {
+          pointToLayer: function (feature, latlng) {
+            return L.marker(latlng, { icon: customSearchPointIconStyle })
+          }
+        })
+        geojsonLayer.setStyle((feature) => {
+          return styleFeatures(feature, geojsonLayer)
+        })
+        geojsonLayer.options.interactive = false
+        layer.addLayer(geojsonLayer)
+        store.dispatch(setsearchGeojsonBoundary(geojson))
+      }
+    })
+  }
+}
+
+export async function parseGeomUpload(geom) {
+  if (GeoJSONValidation.isValidFeatureCollection(geom)) {
+    if (geom.features.length > 1) {
+      throw Error('Only FeatureCollections with a single feature are supported')
+    }
+    return geom.features[0]
+  }
+  if (GeoJSONValidation.isValidFeature(geom)) {
+    return geom
+  }
+  if (GeoJSONValidation.isValidGeometry(geom)) {
+    return {
+      type: 'Feature',
+      geometry: geom,
+      properties: {}
+    }
+  }
+  throw Error('Invalid geojson uploaded')
+}
+
+function styleFeatures(feature, geojsonLayer) {
+  if (
+    feature.geometry.type === 'LineString' ||
+    feature.geometry.type === 'MultiLineString'
+  ) {
+    return customSearchLineStyle
+  }
+  if (
+    feature.geometry.type === 'Polygon' ||
+    feature.geometry.type === 'MultiPolygon'
+  ) {
+    return customSearchPolygonStyle
+  }
+  if (feature.geometry.type === 'GeometryCollection') {
+    const accumulatedStyle = {}
+    feature.geometry.geometries.forEach((part) => {
+      if (part.type === 'LineString' || part.type === 'MultiLineString') {
+        Object.assign(accumulatedStyle, customSearchLineStyle)
+      }
+      if (part.type === 'Polygon' || part.type === 'MultiPolygon') {
+        Object.assign(accumulatedStyle, customSearchPolygonStyle)
+      }
+    })
+    return accumulatedStyle
   }
 }